# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wholesome Uganda API Docs",
    "version": "1.0.0",
    "description": "API documentation for Wholesome Uganda website"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local dev server"
    },
    {
      "url": "https://api.wholesomeug.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/blogs": {
      "get": {
        "summary": "Get all blog posts",
        "tags": [
          "Blog"
        ],
        "responses": {
          "200": {
            "description": "List of blogs"
          }
        }
      },
      "post": {
        "summary": "Create a new blog post",
        "tags": [
          "Blog"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "description": "Blog date (YYYY-MM-DD)"
                  },
                  "category": {
                    "type": "string"
                  },
                  "excerpt": {
                    "type": "string"
                  },
                  "readTime": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog created"
          }
        }
      }
    },
...
    "/contact": {
      "post": {
        "summary": "Submit a contact message",
        "tags": [
          "Contact"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "example": "john@example.com"
                  },
                  "subject": {
                    "type": "string",
                    "example": "Question about programs"
                  },
                  "message": {
                    "type": "string",
                    "example": "Tell me more about your initiatives."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message submitted"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "get": {
        "summary": "Get all contact messages",
        "tags": [
          "Contact"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Messages per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of contact messages"
          }
        }
      }
    },
...
    "/dashboard": {
      "get": {
        "summary": "Get dashboard statistics",
        "tags": [
          "Dashboard"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard data returned"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/gallery": {
      "get": {
        "summary": "Get all gallery items",
        "tags": [
          "Gallery"
        ],
        "responses": {
          "200": {
            "description": "List of gallery items"
          }
        }
      },
      "post": {
        "summary": "Upload multiple gallery items",
        "tags": [
          "Gallery"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "Optional description to apply to all uploaded images"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Array of image files"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Images uploaded successfully"
          },
          "400": {
            "description": "Bad request (e.g. no files or file too large)"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/newsletter": {
      "get": {
        "summary": "Get all subscribers",
        "tags": [
          "Newsletter"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of subscribed emails"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/revalidate": {
      "post": {
        "summary": "Trigger ISR revalidation (single page or all pages)",
        "tags": [
          "Revalidation"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "**Admins only.**\n- **Single-page**: send JSON `{ \"path\": \"/about/people\" }`\n- **All pages**: send an empty body (`{}` or no JSON)\n",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "path": {
                    "type": "string",
                    "example": "/about/people",
                    "description": "Absolute path to revalidate"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Revalidation triggered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Revalidation triggered"
                    },
                    "result": {
                      "type": "object",
                      "description": "Raw response from Next.js ISR endpoint"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid `path` parameter"
          },
          "401": {
            "description": "Unauthorized (missing/invalid JWT)"
          },
          "500": {
            "description": "Server or network error"
          }
        }
      }
    }
  },
...
}